Refactor DropDatabaseDiscoveryRuleStatementUpdaterTest
